RecognizerExporterSetter

+++ grade/internal/domain/endorsed/endorsed_receive_endorsement_test.go
@@ -71,7 +71,7 @@ func TestEndorsedCanCompleteEndorsement(t *testing.T) {
 			return nil
 		}, recognizer.ErrNoEndorsementReservation},
 		{func(r *recognizer.Recognizer) error {
-			for i := uint8(0); i < recognizer.YearlyEndorsementCount; i++ {
+			for i := uint(0); i < recognizer.YearlyEndorsementCount; i++ {
 				err := r.ReserveEndorsement()
 				if err != nil {
 					return err
